+++ packages/kit/src/runtime/server/endpoint.js
@@ -1,9 +1,18 @@
 import { lowercase_keys } from './utils.js';
 
+/** @param {string} body */
+function error(body) {
+	return {
+		status: 500,
+		body,
+		headers: {}
+	};
+}
+
 /**
  * @param {import('types/endpoint').ServerRequest} request
  * @param {import('types/internal').SSREndpoint} route
- * @returns {Promise<import('types/endpoint').ServerResponse>}
+ * @returns {Promise<import('types/hooks').ServerResponse>}
  */
 export default async function render_route(request, route) {
 	const mod = await route.load();
@@ -19,27 +28,40 @@ export default async function render_route(request, route) {
 
 		if (response) {
 			if (typeof response !== 'object') {
-				return {
-					status: 500,
-					body: `Invalid response from route ${request.path};
-						 expected an object, got ${typeof response}`,
-					headers: {}
-				};
+				return error(
+					`Invalid response from route ${request.path}: expected an object, got ${typeof response}`
+				);
 			}
 
 			let { status = 200, body, headers = {} } = response;
 
 			headers = lowercase_keys(headers);
+			const type = headers['content-type'];
 
-			if (
-				typeof body === 'object' &&
-				(!('content-type' in headers) || headers['content-type'] === 'application/json')
-			) {
-				headers = { ...headers, 'content-type': 'application/json' };
-				body = JSON.stringify(body);
+			// validation
+			if (type === 'application/octet-stream' && !(body instanceof Uint8Array)) {
+				return error(
+					`Invalid response from route ${request.path}: body must be an instance of Uint8Array if content type is application/octet-stream`
+				);
 			}
 
-			return { status, body, headers };
+			if (body instanceof Uint8Array && type !== 'application/octet-stream') {
+				return error(
+					`Invalid response from route ${request.path}: Uint8Array body must be accompanied by content-type: application/octet-stream header`
+				);
+			}
+
+			/** @type {string | Uint8Array} */
+			let normalized_body;
+
+			if (typeof body === 'object' && (!type || type === 'application/json')) {
+				headers = { ...headers, 'content-type': 'application/json' };
+				normalized_body = JSON.stringify(body);
+			} else {
+				normalized_body = /** @type {string | Uint8Array} */ (body);
+			}
+
+			return { status, body: normalized_body, headers };
 		}
 	}
 }