+++ packages/adapter-cloudflare-workers/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sveltejs/adapter-cloudflare-workers
 
+## 1.0.0-next.7
+
+### Patch Changes
+
+- 11e7840: Ensure rawBody is a string or Uint8Array
+
 ## 0.0.2-next.6
 
 ### Patch Changes